refactor: matchLottos 로직 개선 및 코드 단순화
