+++ src/main/kotlin/lotto/util/Rank.kt
@@ -1,12 +1,12 @@
 package lotto.util
 
 enum class Rank(val countOfMatch: Int, val price: Int) {
-    FIRST(6, 2_000_000_000),
-    SECOND(5, 30_000_000),
-    THIRD(5, 1_500_000),
-    FOURTH(4, 50_000),
+    NONE(0, 0),
     FIFTH(3, 5_000),
-    NONE(0, 0), ;
+    FOURTH(4, 50_000),
+    THIRD(5, 1_500_000),
+    SECOND(5, 30_000_000),
+    FIRST(6, 2_000_000_000), ;
 
     companion object {
         fun getRankState(
