+++ src/test/kotlin/lotto/domain/LottoCalculatorTest.kt
@@ -4,27 +4,27 @@ import lotto.util.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-class LottoResultTest {
+class LottoCalculatorTest {
     val number: Lotto = Lotto(intToLottoNumber(listOf(1, 2, 3, 4, 5, 6)))
     val winningNumber: Lotto = Lotto(intToLottoNumber(listOf(3, 4, 5, 6, 7, 8)))
 
     @Test
     fun `6개 숫자를 넣으면 일치한 숫자의 개수를 알 수 있다`() {
-        val lottoGame = LottoResult(winningNumber, LottoNumber(1))
+        val lottoGame = LottoCalculator(winningNumber, LottoNumber(1))
 
         assertThat(lottoGame.compareLotto(number.numbers)).isEqualTo(4)
     }
 
     @Test
     fun `보너스 번호가 일치하면 true를 반환한다`() {
-        val lottoGame = LottoResult(winningNumber, LottoNumber(1))
+        val lottoGame = LottoCalculator(winningNumber, LottoNumber(1))
 
         assertThat(lottoGame.checkBonusNumber(number.numbers)).isTrue()
     }
 
     @Test
     fun `보너스 번호가 일치하지 않으면 false를 반환한다`() {
-        val lottoGame = LottoResult(winningNumber, LottoNumber(45))
+        val lottoGame = LottoCalculator(winningNumber, LottoNumber(45))
 
         assertThat(lottoGame.checkBonusNumber(number.numbers)).isFalse()
     }
@@ -39,8 +39,8 @@ class LottoResultTest {
         val winningNumber: Lotto = Lotto(intToLottoNumber(listOf(2, 3, 4, 5, 6, 7)))
         val winningBonusNumber: LottoNumber = LottoNumber(45)
 
-        val lottoGame = LottoResult(winningNumber, winningBonusNumber)
-        val winningStats = lottoGame.matchLotto(lottos)
+        val lottoGame = LottoCalculator(winningNumber, winningBonusNumber)
+        val winningStats = lottoGame.matchLottos(lottos)
 
         val correctWinningStats: Map<Rank, Int> =
             mapOf(
@@ -49,6 +49,7 @@ class LottoResultTest {
                 Rank.THIRD to 1,
                 Rank.FOURTH to 0,
                 Rank.FIFTH to 0,
+                Rank.NONE to 0,
             )
         assertThat(winningStats).isEqualTo(correctWinningStats)
     }
@@ -58,7 +59,7 @@ class LottoResultTest {
         val winningNumber: Lotto = Lotto(intToLottoNumber(listOf(2, 3, 4, 5, 6, 7)))
         val winningBonusNumber: LottoNumber = LottoNumber(45)
 
-        val lottoResult = LottoResult(winningNumber, winningBonusNumber)
+        val lottoResult = LottoCalculator(winningNumber, winningBonusNumber)
 
         val winningStats = mapOf(Rank.FIRST to 1, Rank.SECOND to 1, Rank.THIRD to 1, Rank.FOURTH to 1, Rank.FIFTH to 1)
 
@@ -73,7 +74,7 @@ class LottoResultTest {
         val winningNumber: Lotto = Lotto(intToLottoNumber(listOf(2, 3, 4, 5, 6, 7)))
         val winningBonusNumber: LottoNumber = LottoNumber(45)
 
-        val lottoResult = LottoResult(winningNumber, winningBonusNumber)
+        val lottoResult = LottoCalculator(winningNumber, winningBonusNumber)
 
         assertThat(lottoResult.calculateProfit(totalPrize, purchaseAmount)).isEqualTo(totalPrize / purchaseAmount.toDouble())
     }
